Create Possible Python approach
You are given N numbers you can decrease each number by 10 % or K kg whichever is more. If the number you choose is less than K you can make it zero. Your task is to minimize the sum of the number as small as possible by performing N operations only

Note: Final answer should be typecasted to the nearest integer smaller or equal to the value.